DE-31 Update integration spec
