+++ spec/cassettes/Mailjet_API_Resource/retrieves_all_contact_lists.yml
@@ -1,0 +1,42 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://api.mailjet.com/v3/REST/contactslist?Limit=0
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept:
+      - application/json
+      User-Agent:
+      - mailjet-api-v3-ruby/1.6.0
+      Accept-Encoding:
+      - deflate
+      Content-Type:
+      - application/json
+      Host:
+      - api.mailjet.com
+      Authorization:
+      - "<BASIC>"
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Fri, 06 Nov 2020 09:20:19 GMT
+      Content-Type:
+      - application/json
+      Content-Length:
+      - '340'
+    body:
+      encoding: UTF-8
+      string: '{ "Count" : 2, "Data" : [{ "Address" : "P8zbwo0g3", "CreatedAt" : "2020-09-16T10:35:55Z",
+        "ID" : 32964, "IsDeleted" : false, "Name" : "UnDPbpLmQP", "SubscriberCount"
+        : 1 }, { "Address" : "tyra0hcvz", "CreatedAt" : "2020-09-26T18:48:53Z", "ID"
+        : 33952, "IsDeleted" : false, "Name" : "my_contactslist", "SubscriberCount"
+        : 1 }], "Total" : 2 }'
+    http_version: null
+  recorded_at: Fri, 06 Nov 2020 09:20:19 GMT
+recorded_with: VCR 5.1.0
