+++ spec/cassettes/Mailjet_API_Resource/the_resource_path_does_not_exist/raises_an_error.yml
@@ -1,0 +1,39 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://api.mailjet.com/v3/wrongresourcepath?Limit=0
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept:
+      - application/json
+      User-Agent:
+      - mailjet-api-v3-ruby/1.6.0
+      Accept-Encoding:
+      - deflate
+      Content-Type:
+      - application/json
+      Host:
+      - api.mailjet.com
+      Authorization:
+      - "<BASIC>"
+  response:
+    status:
+      code: 500
+      message: Internal error
+    headers:
+      Date:
+      - Mon, 09 Nov 2020 03:20:03 GMT
+      Content-Type:
+      - application/json
+      Content-Length:
+      - '138'
+    body:
+      encoding: UTF-8
+      string: |
+        { "ErrorInfo" : "EFPWebError", "ErrorMessage" : "Could not determine HTTP module for request \"wrongresourcepath\"", "StatusCode" : 500 }
+    http_version: null
+  recorded_at: Mon, 09 Nov 2020 03:20:03 GMT
+recorded_with: VCR 5.1.0
